src/test6/auth.py
"""Authentication utilities: password hashing and JWT token management."""

import os
import secrets
import string
from datetime import UTC, datetime, timedelta

import bcrypt
from dotenv import load_dotenv
from jose import JWTError, jwt

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY", secrets.token_hex(32))
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24  # 24 hours


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Return True if the plain password matches the stored hash."""
    return bcrypt.checkpw(plain_password.encode(), hashed_password.encode())


def get_password_hash(password: str) -> str:
    """Return a bcrypt hash of the given password."""
    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()


def create_access_token(data: dict[str, str]) -> str:
    """Create and return a signed JWT access token."""
    to_encode: dict = dict(data)
    expire = datetime.now(UTC) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode["exp"] = int(expire.timestamp())
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)  # type: ignore[no-any-return]


def decode_token(token: str) -> str | None:
    """Decode a JWT token and return the subject email, or None if invalid."""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str | None = payload.get("sub")
        return email
    except JWTError:
        return None


def generate_account_number() -> str:
    """Generate a unique random 10-digit account number prefixed with ACC."""
    digits = "".join(secrets.choice(string.digits) for _ in range(10))
    return f"ACC{digits}"
