pyproject.toml
[project]
name = "test6"
version = "0.1.0"
description = "test"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "andy", email = "andy@edgered.com.au" }
]
license = { text = "MIT" }

dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.8",          # keep in sync with .pre-commit-config.yaml ruff-pre-commit rev and super-linter version
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
    "mypy==1.19.0",          # keep in sync with super-linter version
    "detect-secrets==1.5.0", # keep in sync with .pre-commit-config.yaml detect-secrets rev
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle (docstrings)
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]

ignore = [
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D203",    # Conflicts with D211
    "D213",    # Conflicts with D212
    "G004",    # Logging uses f-string
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Use of assert (pytest uses assertions)
    "D",       # No docstrings required in tests
    "PLR2004", # Magic value used in comparison
]
"src/test6/**/*.py" = [
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D107",    # Missing docstring in __init__
    "EM101",   # Exception message in f-string (too pedantic for API handlers)
    "EM102",   # Exception message in f-string
    "TRY003",  # Avoid specifying long messages outside the exception class
    "ARG001",  # Unused function argument (common in FastAPI dependency injection)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["test6"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/"]

[tool.hatch.build.targets.wheel]
packages = ["src/test6"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
